+++ wasm-node/src/node.rs
@@ -1,9 +1,9 @@
 use anyhow::Context;
-use celestia_node::network::network_id;
+use celestia_node::network::{network_id, canonical_network_bootnodes, network_genesis};
 use celestia_node::node::{Node, NodeConfig};
 use celestia_node::store::IndexedDbStore;
 use celestia_types::hash::Hash;
-use js_sys::{Array, JsString};
+use js_sys::Array;
 use libp2p::identity::Keypair;
 use libp2p::{identity, Multiaddr};
 use serde_wasm_bindgen::to_value;
@@ -133,14 +133,24 @@ impl WasmNode {
 #[wasm_bindgen(js_class = NodeConfig)]
 impl WasmNodeConfig {
     #[wasm_bindgen(constructor)]
-    pub fn new(network: Network, genesis_hash: JsString, bootnodes: Vec<JsString>) -> Self {
+    pub fn new(network: Network) -> Self {
         let p2p_local_keypair = identity::Keypair::generate_ed25519();
+        let genesis_hash = network_genesis(network.into());
 
-        let p2p_bootnodes: Vec<Multiaddr> = bootnodes
-            .iter()
-            .map(|addr| addr.as_string().unwrap_throw().parse().unwrap_throw())
-            .collect::<Vec<_>>();
-        let genesis_hash = genesis_hash.as_string().map(|v| v.parse().unwrap_throw());
+        let mut p2p_bootnodes = canonical_network_bootnodes(network.into());
+        if network == Network::Mocha {
+            // 40.85.94.176 is a node set up for testing QUIC/WebTransport since official nodes
+            // don't have that enabled currently
+            let webtransport_bootnode_addrs = [
+                "/ip4/40.85.94.176/tcp/2121/p2p/12D3KooWNJ3Nf1DTQTz8JZogg2eSvPKKKv8itC6fxxspe4C6bizs",
+                "/ip4/40.85.94.176/udp/2121/quic-v1/p2p/12D3KooWNJ3Nf1DTQTz8JZogg2eSvPKKKv8itC6fxxspe4C6bizs",
+                "/ip4/40.85.94.176/udp/2121/quic-v1/webtransport/certhash/uEiBf-OX4HzFK9owOpjdCifsDIWRO0SoD3j3vGKlq0pAXKw/certhash/uEiCx1md1BATJ_0NXAjp3KOuwRYG1535E7kUzFdMq8aPaWw/p2p/12D3KooWNJ3Nf1DTQTz8JZogg2eSvPKKKv8itC6fxxspe4C6bizs",
+                "/ip4/40.85.94.176/udp/2121/quic-v1/p2p/12D3KooWQUYAApYb4DJnhS1QmAwRr5HRvUeHJYocchCpwEhCtDGu",
+                "/ip4/40.85.94.176/udp/2121/quic-v1/webtransport/certhash/uEiBr4-sr95BpqfA-ttpjiLdjbGABhTvX8oxrTXf3Ubfibw/certhash/uEiBSVgyze9xG1UbbNuTwyEUWLPq7l2N9pyeQSs3OtEhGRg/p2p/12D3KooWQUYAApYb4DJnhS1QmAwRr5HRvUeHJYocchCpwEhCtDGu",
+            ].iter().map(|s| s.parse().unwrap_throw());
+
+            p2p_bootnodes.extend(webtransport_bootnode_addrs);
+        }
 
         WasmNodeConfig {
             network,
